moved userRequest Error to seperate folder
